requirements.txt
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
fastapi==0.94.0
grep==0.3.2
h11==0.14.0
hiredis==2.2.2
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==8.14.0
order==2.1.1
packaging==23.0
pluggy==1.0.0
pptree==3.1
pycparser==2.21
pydantic==1.10.6
pytest==7.2.2
pytest-order==1.1.0
python-dotenv==1.0.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.5.1
redis-om==0.1.2
rfc3986==1.5.0
scinum==1.4.4
six==1.16.0
sniffio==1.3.0
starlette==0.26.0.post1
types-pyOpenSSL==23.0.0.4
types-redis==4.5.1.5
typing_extensions==4.5.0
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
